Include the applicable exclusions when running upstream tests
This adds support in the `npm run test` command when running upstream
tests so that any test filters file that can be applied to the current
test suite will be applied to avoid certain tests from running.

This means that `npm run test` will automatically add the relevant
--test-launcher-filter-file parameter if at least one of the following
files are present on-disk at the time of running the tests:

  //brave/test/filters/<test suite>.filter
  //brave/test/filters/<test suite>-<OS>.filter
  //brave/test/filters/<test suite>-<OS>-arch.filter

In case multiple files are available, all of them will be used along
with the --test-launcher-filter-file parameter, so that we can provide
a base filters file for all platforms and then refine it with more
exclusions for specific platforms and architectures.